Define size on monoidal categories instead of arbitrary categories
